+++ Amethyst/AST/MethodNode.cs
@@ -13,6 +13,8 @@ namespace Amethyst.AST
 			Parameters.Insert(0, new AbstractParameter(ParameterModifiers.Macro, new AbstractReferenceTypeSpecifier(Location, new SimpleAbstractTypeSpecifier(Location, string.Join('/', ID.ToString().Split('/')[..^1]))), "this"));
 			base.Process(ctx, root);
 		}
+
+		protected override NamespacedID Mangle(TypeArray args) => base.Mangle(new(args.Types.Skip(1)));
 	}
 
 	public class ConstructorNode(LocationRange loc, List<NamespacedID> tags, FunctionModifiers modifiers, NamespacedID id, List<AbstractParameter> parameters, Expression? baseCall, BlockNode body) : MethodNode(loc, tags, modifiers, new SimpleAbstractTypeSpecifier(loc, id.ToString()), id, parameters, body)
@@ -29,5 +31,7 @@ namespace Amethyst.AST
 			constructor.Body.Add(new ReturnStatement(Location, new VariableExpression(Location, "this")));
 			constructor.Process(ctx, root);
 		}
+
+		protected override NamespacedID Mangle(TypeArray args) => args.Mangle(ID);
 	}
 }