+++ Geode/GeodeBuilder.cs
@@ -205,7 +205,7 @@ namespace Geode
 
 		public (FunctionContext ctx, FunctionValue func) AnonymousFunction(FunctionType type)
 		{
-			var func = new FunctionValue(new("amethyst", InternalPath + "/" + RandomString), type);
+			var func = new FunctionValue(new("amethyst", InternalPath + "/" + RandomString), type, LocationRange.None);
 			var ctx = new FunctionContext(Compiler, func, [], LocationRange.None);
 			AddFunctions(ctx);
 			return (ctx, func);