+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 * Added ranges.
 * Added support for more target selector options.
 * Changed the default rendering prefix for doubles to nothing.
+* Added function overloading.
+*